Move /api/config API controllers into separate module for code modularity
